Improve exchangeHalos() efficiency by an order of magnitude...
While the previous code was short to write,
it was actually very difficult to understand
the logic execution of the code.

This change introduces a few lines of code,
but greatly improves the clarity of intention
in the process being carried out.

Additionally, the previous code resulted in a
runtime *cascade* of MPI sending where each
rightward message would only be sent after the
previous rightward message was received.

This scales linearly in the number of threads
participating in the cluster. This does not deliver
the parallelism that we intuitively intend for
a distributed mechanism like MPI.
The new system trades a few lines of code for
a single, parallel single-directional
transfer operation which is an
order of magnitude for efficient.

This also renames the tags to better align with
their emphasis on directionality.